# Perf testing suite for Awfy