--- a/Gemfile
+++ b/Gemfile
@@ -5,11 +5,20 @@ source "https://rubygems.org"
 gemspec
 
 gem "quickdraw", git: "https://github.com/joeldrapper/quickdraw.git"
-gem "benchmark-ips"
 
 group :development do
 	gem "solargraph", platform: :ruby
 	gem "rubocop", platform: :ruby
 	gem "ruby-lsp", platform: :ruby
 	gem "simplecov", platform: :ruby
+
+	# Benchmarking
+	gem "awfy"
+	# Gems to compare to:
+	gem "activemodel"
+	gem "dry-initializer"
+	gem "dry-types"
+	gem "dry-struct"
+	gem "ruby-enum"
+	gem "typesafe_enum"
 end