Uses https://github.com/stevegeek/awfy

Tests in `./benchmarks/`

Eg comparing `Struct.new` for various implementations with YJIT both enabled and disabled, and across two branches, "main" and "perf"

```
bundle exec awfy ips Struct ".new" --compare-with=main
Running IPS for:
> Struct/.new...
> Report (mri - branch 'perf'): .new
> Report (yjit - branch 'perf'): .new
> Report (mri - branch 'main'): .new
> Report (yjit - branch 'main'): .new
+-----------------------------------------------------------------------------------+
|                              Summary for Struct/.new                              |
+--------+---------+----------------------------+---------+-------------------------+
| Branch | Runtime | Name                       | IPS     | Diff v baseline (times) |
+--------+---------+----------------------------+---------+-------------------------+
| perf   | yjit    |                  Ruby Data | 3598962 |                  2.41 x |
| perf   | mri     |                  Ruby Data | 3484919 |                  2.34 x |
| perf   | yjit    |                Ruby Struct | 3523499 |                  2.36 x |
| perf   | mri     |                Ruby Struct | 3348478 |                  2.25 x |
| perf   | yjit    |            Literal::Struct | 2534041 |                   1.7 x |
| main   | yjit    |            Literal::Struct | 2498048 |                  1.68 x |
| perf   | mri     | Literal::Struct (baseline) | 1491218 |                       - |
| main   | mri     |            Literal::Struct | 1487292 |                same-ish |
| perf   | yjit    |                Dry::Struct | 1304467 |                 -1.14 x |
| perf   | mri     |                Dry::Struct |  634954 |                 -2.35 x |
| perf   | yjit    |    ActiveModel::Attributes |  509929 |                 -2.92 x |
| perf   | mri     |    ActiveModel::Attributes |  316039 |                 -4.72 x |
+--------+---------+----------------------------+---------+-------------------------+
```