@stevegeek 2024-10-30
I've already updated the branch to just use Awfy. I'll reword PR. Hopefully will add more to the benchmark suite soon. We could also house this in a separate repo if you want